+++ src/phpDebug.ts
@@ -219,6 +219,7 @@ class PhpDebugSession extends vscode.DebugSession {
                 },
             ],
             supportTerminateDebuggee: true,
+            supportsDelayedStackTraceLoading: false,
         }
         this.sendResponse(response)
     }
@@ -741,7 +742,7 @@ class PhpDebugSession extends vscode.DebugSession {
             if (!connection) {
                 throw new Error('Unknown thread ID')
             }
-            const { stack } = await connection.sendStackGetCommand()
+            let { stack } = await connection.sendStackGetCommand()
             // First delete the old stack trace info ???
             // this._stackFrames.clear();
             // this._properties.clear();
@@ -781,7 +782,10 @@ class PhpDebugSession extends vscode.DebugSession {
                 this._errorStackFrames.set(id, status)
                 response.body = { stackFrames: [{ id, name, source, line, column: 1 }] }
             } else {
+                const totalFrames = stack.length
+                stack = stack.slice(args.startFrame, args.levels ? (args.startFrame ?? 0) + args.levels : undefined)
                 response.body = {
+                    totalFrames,
                     stackFrames: stack.map((stackFrame): VSCodeDebugProtocol.StackFrame => {
                         let source: VSCodeDebugProtocol.Source
                         let line = stackFrame.line