+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.25.0]
+
+- Implement delayed stack loading with startFrame and levels argument to StackTrace Request
+
 ## [1.24.3]
 
-- Fox for broken property traversal #755
+- Fix for broken property traversal #755
 
 ## [1.24.2]